Fix cursor pixel trail position
Because the grid is centered in the page, and the pixel trail was responding as if the grid was at the top left side of the page, the pixel trail was being generated below the cursor. Now, by modifying the js to take the grid's position into account, the cursor lines up perfectly.