Fix oldGrid removal
As of before, if the number of boxes inputed was greater than zero, the calculation did not actually remove the entire grid. In order to fix this, the replaceGrid mechanic has been redeclared as a function and it uses  a simple while loop to delete every box when a new number of boxes is input.